added auth and email sending

+++ controllers/admin.js
@@ -30,6 +30,9 @@ exports.getEditProduct = async (req, res) => {
   const id = req.params.productId;
   try {
     const product = await Product.findById(id);
+    if (product.userId.toString() !== req.user._id.toString()) {
+      return res.redirect("/");
+    }
     if (!product) throw new Error("No product");
     res.render("admin/edit-product", {
       pageTitle: "Edit Product",
@@ -53,6 +56,9 @@ exports.postEditProduct = async (req, res) => {
   try {
     const product = await Product.findById(newData.id);
     if (!product) throw new Error(product);
+    if (product.userId.toString() !== req.user._id.toString()) {
+      return res.redirect("/");
+    }
     product.title = newData.title;
     product.imageUrl = newData.imageUrl;
     product.price = newData.price;
@@ -70,6 +76,9 @@ exports.postEditProduct = async (req, res) => {
 exports.getDeleteProduct = async (req, res) => {
   const id = req.params.productId;
   try {
+    const product = await Product.findById(id);
+    if (!product || product.userId.toString() !== req.user._id.toString())
+      return res.redirect("/");
     const result = await Product.findByIdAndRemove(id);
     if (!result) throw new Error("Product not deleted");
     res.redirect("/admin/products");
@@ -80,7 +89,7 @@ exports.getDeleteProduct = async (req, res) => {
 
 exports.getProducts = async (req, res) => {
   try {
-    const products = await Product.find();
+    const products = await Product.find({ userId: req.user._id });
     if (!products) throw new Error("Products was not finded");
     res.render("admin/product-list", {
       pageTitle: "Products",
